+++ tests/EngineWorks/Templates/Utils.php
@@ -1,0 +1,10 @@
+<?php
+namespace Tests\EngineWorks\Templates;
+
+class Utils
+{
+    public static function samples($append = '')
+    {
+        return dirname(dirname(__DIR__)) . '/samples/' . $append;
+    }
+}