+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# version 2.0.0
+- Create `Resolver` class to retrieve the location of the template based on a template name
+    - This allow to call `fetch` inside a `Template object` to include other template
+    - This allow to implement different logic to retrieve the template file location
+- `Templates` remove the logic of resolver, add logic for `defaultResolver`
+- `Template` allows to set a resolver, if none provided it creates a Resolver object
+- Include dependences to upload coverage to scrutinizer and coveralls
+- This breaks compatibility with previous versions
+
 # version 1.0.1
 - Looks like I should not test readable against '/dev/console', travis consider the file as readable
 - Improve the test to avoid build failure